dojson: move force_force_list to inspirehep.utils
* Closes #1326.
