+++ inspirehep/dojson/hep/fields/bd5xx.py
@@ -27,7 +27,9 @@ from __future__ import absolute_import, division, print_function
 from dojson import utils
 
 from ..model import hep, hep2marc
-from ...utils import force_force_list, get_record_ref
+from ...utils import get_record_ref
+
+from inspirehep.utils.helpers import force_force_list
 
 
 @hep.over('public_notes', '^500..')
