+++ inspirehep/dojson/hep/model.py
@@ -27,7 +27,9 @@ from __future__ import absolute_import, division, print_function
 from dojson import Overdo
 
 from ..schema import SchemaOverdo
-from ..utils import force_force_list, get_record_ref
+from ..utils import get_record_ref
+
+from inspirehep.utils.helpers import force_force_list
 
 
 def add_book_info(record, blob):
