+++ inspirehep/dojson/utils/__init__.py
@@ -38,6 +38,7 @@ from inspirehep.config import (
 )
 
 from inspirehep.utils.dedupers import dedupe_list, dedupe_list_of_dicts
+from inspirehep.utils.helpers import force_force_list
 
 
 def classify_field(value):
@@ -89,21 +90,6 @@ def create_profile_url(profile_id):
         return base_url + str(profile_id)
     except (TypeError, ValueError):
         return ''
-
-
-def force_force_list(data):
-    """Wrap data in list.
-
-    We need to define this awkardly named method because DoJSON's method
-    force_list returns tuples or None instead of lists.
-    """
-    if data is None:
-        return []
-    elif not isinstance(data, (list, tuple, set)):
-        return [data]
-    elif isinstance(data, (tuple, set)):
-        return list(data)
-    return data
 
 
 def force_single_element(obj):
